Add pseudo.type.operation
