+++ pseudo/type/__init__.py
@@ -12,103 +12,6 @@ from pseudo.type.variable import Variable, Assignment
 from pseudo.type.base import Value, EOL
 
 
-GROUP_1 = {"*", "div", "mod"}
-GROUP_2 = {"-", "+"}
-
-
-class Operator(Value):
-    """Opartor class for representing mathematical operator."""
-
-    def eval(self, left: Value, right: Value):
-        if self.value == "+":
-            return left.eval() + right.eval()
-        if self.value == "-":
-            return left.eval() - right.eval()
-        if self.value == "*":
-            return left.eval() * right.eval()
-        if self.value == "div":
-            return left.eval() // right.eval()
-        if self.value == "mod":
-            return left.eval() % right.eval()
-
-        if self.value == "=":
-            if left.eval() == right.eval():
-                return 1
-            return 0
-        if self.value == "!=":
-            if left.eval() != right.eval():
-                return 1
-            return 0
-        if self.value == ">":
-            if left.eval() > right.eval():
-                return 1
-            return 0
-        if self.value == "<":
-            if left.eval() < right.eval():
-                return 1
-            return 0
-        if self.value == "<=":
-            if left.eval() <= right.eval():
-                return 1
-            return 0
-        if self.value == ">=":
-            if left.eval() >= right.eval():
-                return 1
-            return 0
-
-    def __lt__(self, o):
-        if self.value in GROUP_2:
-            if o.value in GROUP_1:
-                return True
-        return False
-
-    def __gt__(self, o):
-        if self.value in GROUP_1:
-            return True
-        if self == o:
-            return True
-        return False
-
-    def __eq__(self, o):
-        try:
-            if self.value == o.value:
-                return True
-            if self.value in GROUP_1 and o.value in GROUP_1:
-                return True
-            if self.value in GROUP_2 and o.value in GROUP_2:
-                return True
-            return False
-        except AttributeError:
-            return False
-
-    def __str__(self):
-        return self.__repr__()
-
-    def __repr__(self):
-        return f"Operator({repr(self.value)})"
-
-
-class Operation:
-    """Operation node."""
-
-    def __init__(self, operator, left, right):
-        self.operator = operator
-        self.left = left
-        self.right = right
-
-    def eval(self):
-        return self.operator.eval(self.left, self.right)
-
-    def __eq__(self, other):
-        try:
-            return self.__dict__ == other.__dict__
-        except AttributeError:
-            return False
-
-    def __repr__(self):
-        return f"Operation({self.operator}, {self.left}, {self.right})"
-
-
 class Statement:
     """
     Node for statement with arguments.
